+++ inst/plumber/v1/endpoints.R
@@ -538,3 +538,13 @@ function(req) {
   params$version <- NULL
   do.call(pipapi::ui_svy_meta, params)
 }
+
+#* Return valid years
+#* @get /api/v1/valid-years
+#* @param data_dir:[chr] Path to the main data directory
+#* @serializer json list(na="null")
+function(req) {
+  params <- req$argsQuery
+  params$lkup <- lkups$versions_paths[[params$version]]
+  pipapi::valid_years(data_dir = params$lkup$data_root)
+}